feat: Add task to show configured values
The plugins gets properties from a bunch of places, from environment
variables, from global configuration files, from gradle properties. On
top off that the configuration files can be configured with environment
variables. All of this makes it hard to know if something failed because
the configuration is wrong, or for some other reason.

Add a task that will list configuration properties as seen by this
plugin.

Some of the configuration properties contain passwords, these are hidden
with astirixes. If you do want to show the passwords you need to set
``-PshowPasswords=true``