feat: Use beta nexus repository for plugin
The plugin makes a distinction between releases and beta releases for
the projects that are build using it. The plugin project itself however
did not have any such notion.

Add a check to see if the git generated version contains a hash number.
If that is the case we use the ``maven-pre-releases`` repository instead
of ``maven-releases``.

We still check if the version indicates a dirty build. Such builds (if
published) would go to ``maven-dirties`` instead of ``maven-releases``.