refactor: Use elvis operator to find properties
We used to check if a property is present on the project and if so use
it, otherwise provide some default

.. code-block:: groovy

    project.hasProperty ('key') ? property.get ('key') : 'default'

But there exists an operator that can do this sort of check and provided
a default: The Elvis Operator.

We have to use ``findProperty ()`` instead of ``getProperty ()``. The
latter throws an exception when not found, the former will return null.

.. code-block:: groovy

    project.findProperty ('key') ?: 'default'