feat: Only configure the credentials of they are set
When we configured the repository to which we publishing artifacts we
always configured the credentials. We now only configure the credentials
when they are set.

The username gets a default, so that will always be present. The
password however is emtpy by default. If it is empty we don't set any
credentials.