feat: Support credentials on dependency nexus.
The nexus repository used to retrieve dependencies never had any
credentials set. This means that annonymous access has to be enabled on
the nexus in order for projects that use this plugin to success.

You could also provide custom repository definitions in the projects
themselves, but that would defeat the purpose of defining it in the
plugin.

We now check if a password is provided for the base repository, if it is
credentials are provided when we try to retrieve dependencies from the
nexus repository.