feat: Set the test framework for the tests
In later versions of gradle (9.x and up) the automatic detection of
testing framework implementations is no longer supported. So the
framework that is always used can be added by the plugin.

This needs to be done for the default test configuration, but also for
the custom it test configuration.

https://docs.gradle.org/8.2.1/userguide/upgrading_version_8.html#test_framework_implementation_dependencies